Delete unness. .env file
For anyone curious: this `.env` doesn't contain API or SSH keys - it simply contains paths to other git projects on my local machine. This is a file the user of this script must set up for themselves, and in alpha devs of the project it got committed.

Removing it here so that it doesn't confuse anyone. Please, feel free to try and steal my secrets from 4 arbitrary file paths...